// src/app/shared/models/settingshop.ts
export class Settingshop {
    id: number;
    transportMethod: string = '';
    paymentMethod: string = '';
    freeShip: number = 0;
    defaultShip: number = 0;
    noithanh: string = '';
    ngoaithanh: string = '';
    gianoithanh: number = 0;
    giangoaithanh: number = 0;
    chanhxe: number = 0;
    thuho: number = 0;
}